# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Esmalteria Sistema Gerencial - API",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5500/"
    }
  ],
  "paths": {
    "/login/": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Loga o usuario e retorna um token de autenticacao.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login/logout": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Remove o token de autenticacao do usuario.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login/usuario": {
      "get": {
        "tags": [
          "Login"
        ],
        "summary": "Retorna o usuario logado a partir do token.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/usuario/": {
      "get": {
        "tags": [
          "Usuário"
        ],
        "summary": "Lista todos os usuários.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/usuario/desativar/{id}": {
      "post": {
        "tags": [
          "Usuário"
        ],
        "summary": "Desativa um usuário específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/usuario/ativar/{id}": {
      "post": {
        "tags": [
          "Usuário"
        ],
        "summary": "Ativa um usuário específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/profissional/": {
      "post": {
        "tags": [
          "Profissional"
        ],
        "summary": "Cadastra um Profissional.",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/profissional"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profissional"
        ],
        "summary": "Altera os dados de um Profissional.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/alterarProfissional"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Profissional"
        ],
        "summary": "Lista todos os Profissionals.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/profissional/{id}": {
      "get": {
        "tags": [
          "Profissional"
        ],
        "summary": "Obtém os dados de um Profissional específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Profissional"
        ],
        "summary": "Demite um Profissional Específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/cliente/": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "summary": "Cadastra um Cliente.",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/cliente"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cliente"
        ],
        "summary": "Altera os dados de um Cliente.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/alterarCliente"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Lista todos os Clientes.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/cliente/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Obtém os dados de um Cliente específico.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/servico/": {
      "post": {
        "tags": [
          "Serviço"
        ],
        "summary": "Cadastra um Serviço.",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/servico"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Serviço"
        ],
        "summary": "Altera os dados de um Serviço.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/alterarServico"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Serviço"
        ],
        "summary": "Lista todos os Serviços.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/servico/{id}": {
      "get": {
        "tags": [
          "Serviço"
        ],
        "summary": "Busca um Serviço pelo ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Serviço"
        ],
        "summary": "Exclui um Serviço pelo ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/uploads/clientes/{image}": {
      "get": {
        "tags": [
          "Uploads"
        ],
        "summary": "busca a imagem de um Cliente.",
        "description": "",
        "parameters": [
          {
            "name": "image",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "erro": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      },
      "servico": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Manicure",
            "required": true
          },
          "descricao": {
            "type": "string",
            "example": "Detalhes do serviço de manicure",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "alterarServico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "required": true
          },
          "nome": {
            "type": "string",
            "example": "Manicure",
            "required": true
          },
          "descricao": {
            "type": "string",
            "example": "Detalhes do serviço de manicure",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "cliente": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Machado de Assis",
            "required": true
          },
          "telefone": {
            "type": "string",
            "example": "(11) 12345-6789",
            "required": true
          },
          "email": {
            "type": "string",
            "example": "machado@exemplo.com",
            "required": true
          },
          "senha": {
            "type": "string",
            "example": "Senha123@",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "alterarCliente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "required": true
          },
          "telefone": {
            "type": "string",
            "example": "(11) 12345-6789",
            "required": true
          },
          "email": {
            "type": "string",
            "example": "machado@exemplo.com",
            "required": true
          },
          "senha": {
            "type": "string",
            "example": "Senha123@",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "profissional": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Albert Camus",
            "required": true
          },
          "telefone": {
            "type": "string",
            "example": "(11) 98765-4321",
            "required": true
          },
          "email": {
            "type": "string",
            "example": "albert@exemplo.com",
            "required": true
          },
          "senha": {
            "type": "string",
            "example": "Senha123@",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          },
          "cpf": {
            "type": "string",
            "example": "123.456.789-00",
            "required": true
          }
        }
      },
      "alterarProfissional": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "required": true
          },
          "telefone": {
            "type": "string",
            "example": "(11) 98765-4321",
            "required": true
          },
          "email": {
            "type": "string",
            "example": "albert@exemplo.com",
            "required": true
          },
          "senha": {
            "type": "string",
            "example": "Senha123@",
            "required": true
          },
          "foto": {
            "type": "string",
            "format": "binary"
          }
        }
      }
    }
  }
}